Move tabs files to be more idiomatic with the rest of the app
