--- src/pages/task/taskTabs/BuildBaron.ts
@@ -1,5 +1,0 @@
-import { loadable } from "components/SpruceLoader";
-
-export const BuildBaron = loadable(
-  () => import("./buildBaronAndAnnotations/BuildBaron")
-);
